requirements.txt
pandas==1.5.3
streamlit==1.22.0
numpy==1.24.3
Pillow==9.4.0
xlsxwriter==3.0.7
